Correctly align lyrics containing bidirectional text (lyrics page)
This commit updates the live lyrics page to correctly horzintally-align
each `lyricsLine` according to the "primary language" being used. This
superior output is achieved by removing the per-`lyricsLine` `<bdi>`
elements and instead applying `dir="auto"` to the parent
`lyricsContainer` instead. This completely isolates the text direction
of the `lyricsContainer` from the rest of the Jellyfin UI, ensuring that
its contents always look correct regardless of the directionality of the
rest of the interface.

See below for more info on the global `dir=` HTML attribute:

https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Global_attributes/dir